requirements.txt
asgiref==3.3.4
astroid==2.5.3
autopep8==1.5.6
Django==3.2
djangorestframework==3.12.4
flake8==3.9.0
isort==5.8.0
jedi==0.17.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
parso==0.7.1
Pillow==8.2.0
pluggy==0.13.1
pycodestyle==2.6.0
pydocstyle==6.0.0
pyflakes==2.2.0
pylint==2.7.4
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
rope==0.18.0
six==1.15.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
ujson==4.0.2
wrapt==1.12.1
yapf==0.31.0
